Add user page in Admin

+++ MovieTracker.Services/UserService.cs
@@ -36,6 +36,12 @@ namespace MovieTracker.Services
         {
             var user = await _userManager.FindByIdAsync(userId);
             if (user == null) return false;
+
+            if (await _userManager.IsInRoleAsync(user, "Administrator"))
+            {
+                return false;
+            }
+
             var result = await _userManager.DeleteAsync(user);
             return result.Succeeded;
         }
